requirements.txt
yt-dlp>=2024.0.0
google-api-python-client>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
beautifulsoup4>=4.12.0
requests>=2.31.0
pyyaml>=6.0
